requirements.txt
aioredis==1.3.1
amqp==5.0.9
asgiref==3.5.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
black==22.1.0
boto3==1.21.13
botocore==1.24.13
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.11
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dj-rest-auth==2.2.2
Django==4.0.3
django-allauth==0.47.0
django-appconf==1.0.5
django-ckeditor==6.4.1
django-cors-headers==3.11.0
django-debug-toolbar==3.4.0
django-extensions==3.1.5
django-filter==21.1
django-imagekit==4.1.0
django-js-asset==2.0.0
django-phonenumber-field==6.1.0
django-sendgrid-v5==1.2.0
django-silk==4.3.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.21.2
factory-boy==3.2.1
Faker==12.0.1
faker-e164==0.1.0
flake8==4.0.1
future==0.18.2
googlemaps==4.6.0
gprof2dot==2021.2.21
gunicorn==20.1.0
h3==3.7.4
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
incremental==21.3.0
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
jsonfield==3.1.0
jsonschema==4.4.0
kombu==5.2.3
mailchimp-marketing==3.0.74
Markdown==3.3.6
MarkupSafe==2.0.1
mccabe==0.6.1
mixer==7.2.2
msgpack==1.0.3
mypy-extensions==0.4.3
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==21.3
pathspec==0.9.0
phonenumberslite==8.12.45
pilkit==2.0
Pillow==9.0.1
platformdirs==2.4.1
pluggy==1.0.0
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==7.0.1
pytest-django==4.5.2
pytest-factoryboy==2.1.0
python-dateutil==2.8.2
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
redis==4.1.4
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.5.2
sendgrid==6.9.7
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
stripe==2.67.0
timeago==1.0.15
toml==0.10.2
tomli==2.0.0
twilio==7.7.1
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zipp==3.7.0
zope.interface==5.4.0
